+++ TDEngine2/include/ecs/IWorld.h
@@ -118,12 +118,12 @@ namespace TDEngine2
 				Note that the entity won't be deleted, it will be
 				reused later, so a pointer will be valid.
 
-				\param[in] pEntity A pointer to an entity
+				\param[in] entityId An identifier of existing entity
 
 				\return RC_OK if everything went ok, or some other code, which describes an error
 			*/
 
-			TDE2_API virtual E_RESULT_CODE Destroy(CEntity* pEntity) = 0;
+			TDE2_API virtual E_RESULT_CODE Destroy(TEntityId entityId) = 0;
 
 			/*!
 				\brief The method registers specified resource factory within a manager