+++ CHANGELOG.md
@@ -31,6 +31,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - The APIs of **CEntityManager** and **IWorld** were changed. DestroyAllImediately, DestroyImmediately methods were removed.
 
+- The signature of **IWorld::Destroy** was changed.
+
 ### Fixed
 
 - The serialization/deserialization logic for prefabs links was refactored and fixed.